backend/services/ml_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
redis==5.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
xgboost==2.0.3
joblib==1.3.2
numpy==1.24.3
scikit-learn==1.3.2